rename file to keep import paths
