+++ pennylane/ops/qubit/non_parametric_ops.py
